code/requirements.txt
MedPy==0.4.0
numpy==1.21.5
Pillow==10.2.0
scikit_learn==1.0.2
scipy==1.7.3
tensorboardX==2.2
torch==1.11.0
